% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_interface.R
\name{evaluate_gradient_log_likelihoods_all_parameters}
\alias{evaluate_gradient_log_likelihoods_all_parameters}
\title{Evaluate the log of the factors in the model.}
\usage{
evaluate_gradient_log_likelihoods_all_parameters(
  model,
  variables,
  parameters,
  data,
  factor_index = NULL
)
}
\arguments{
\item{model}{The output of calling ilike::parse_ilike_model on a file.}

\item{variables}{The name of the variables to take the derivative with respect to.}

\item{parameters}{A list containing the parameters.}

\item{data}{A list containing the data.}

\item{factor_index}{(optional) The indices giving which factors to evaluate (default is to evaluate all factors in the model).}
}
\value{
A matrix giving the gradient.
}
\description{
Evaluate the log of the factors in the model.
}
