% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_interface.R
\name{evaluate_log_factors}
\alias{evaluate_log_factors}
\title{Evaluate the log of the factors in the model.}
\usage{
evaluate_log_factors(model, parameters, data, factor_index = NULL)
}
\arguments{
\item{model}{The output of calling ilike::compile on a file.}

\item{parameters}{A list containing the parameters.}

\item{data}{A list containing the data.}

\item{factor_index}{(optional) The indices giving which factors to evaluate (default is to evaluate all factors in the model).}
}
\value{
A number giving the log probability,
}
\description{
Evaluate the log of the factors in the model.
}
