% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{get_data}
\alias{get_data}
\title{Load the data from the model.}
\usage{
get_data(model)
}
\arguments{
\item{model}{The output of calling ilike::parse_ilike_model on a file.}
}
\value{
A list containing the data.
}
\description{
Load the data from the model.
}
